Trait lolrofl::model::section::SectionCore

pub trait SectionCore {
    const KIND: u8;

    fn core_len(&self) -> usize;
    fn data_len(&self) -> usize;

    fn kind(&self) -> u8 { ... }
    fn len(&self) -> usize { ... }
    fn raw_data(&self) -> Option<&[u8]> { ... }
}

A generic interface for data segments’ sections

This trait will probably be renamed in future versions

Required Associated Constants

const KIND: u8

The supported section’s ID

This will probably be deprecated in future releases

Required Methods

fn core_len(&self) -> usize

Get the length of the constant part of the section

fn data_len(&self) -> usize

Get the length of the variable part of the section

Provided Methods

fn kind(&self) -> u8

Get the supported section’s ID

This will probably be deprecated in future releases

fn len(&self) -> usize

Get the full length of the section

fn raw_data(&self) -> Option<&[u8]>

Get the raw variable part of the section if any and supported